requirements.txt
apache-airflow-providers-snowflake==5.3.0
apache-airflow-providers-amazon==8.7.1
apache-airflow-providers-slack==8.3.0
faker==37.1.0